Update run_DE_analysis.pl
currently the columns in the output from the following two line do not match up. They are shifted left. Adding col.names = NA will add a "blank" for col 1 which is the gene names column, this will make LofFC and pvalue etc .. match up. I hope this doesnt break the rest of the code. :  
    print $ofh "write.table(result_table, file=\'$output_prefix.edgeR.DE_results\', sep='\t', quote=F, row.names=T)\n";
    print $ofh "write.table(rnaseqMatrix, file=\'$output_prefix.edgeR.count_matrix\', sep='\t', quote=F, row.names=T)\n";